+++ lib/twilio/regions.ts
@@ -27,7 +27,7 @@ export enum Edge {
    * Public edges
    */
   Sydney = 'sydney',
-  SaoPaolo = 'sao-paolo',
+  SaoPaulo = 'sao-paulo',
   Dublin = 'dublin',
   Frankfurt = 'frankfurt',
   Tokyo = 'tokyo',
@@ -149,7 +149,7 @@ const regionURIs: Record<Region, string> = {
  */
 export const edgeToRegion: Record<Edge, Region> = {
   [Edge.Sydney]: Region.Au1,
-  [Edge.SaoPaolo]: Region.Br1,
+  [Edge.SaoPaulo]: Region.Br1,
   [Edge.Dublin]: Region.Ie1,
   [Edge.Frankfurt]: Region.De1,
   [Edge.Tokyo]: Region.Jp1,
@@ -174,7 +174,7 @@ export const edgeToRegion: Record<Edge, Region> = {
  */
 export const regionToEdge: Record<Region, Edge> = {
   [Region.Au1]: Edge.Sydney,
-  [Region.Br1]: Edge.SaoPaolo,
+  [Region.Br1]: Edge.SaoPaulo,
   [Region.Ie1]: Edge.Dublin,
   [Region.De1]: Edge.Frankfurt,
   [Region.Jp1]: Edge.Tokyo,